fix: get formula error
