+++ packages/sheets-formula-ui/src/controllers/formula-alert-render.controller.ts
@@ -17,11 +17,10 @@
 import type { Workbook } from '@univerjs/core';
 import type { IRenderContext, IRenderModule } from '@univerjs/engine-render';
 import { Disposable, Inject, isICellData, LocaleService } from '@univerjs/core';
-import { ErrorType, FormulaDataModel } from '@univerjs/engine-formula';
+import { ErrorType, extractFormulaError, FormulaDataModel } from '@univerjs/engine-formula';
 import { CellAlertManagerService, CellAlertType, HoverManagerService } from '@univerjs/sheets-ui';
 import { IZenZoneService } from '@univerjs/ui';
 import { debounceTime } from 'rxjs';
-import { extractFormulaError } from './utils/utils';
 
 const ALERT_KEY = 'SHEET_FORMULA_ALERT';
