+++ packages/sheets-formula-ui/src/controllers/utils/__tests__/utils.spec.ts
@@ -16,39 +16,10 @@
 
 import { CellValueType } from '@univerjs/core';
 
-import { ErrorType } from '@univerjs/engine-formula';
 import { describe, expect, it } from 'vitest';
-import { extractFormulaError, extractFormulaNumber } from '../utils';
+import { extractFormulaNumber } from '../utils';
 
 describe('Test utils', () => {
-    it('Function extractFormulaError', () => {
-        expect(extractFormulaError({ v: ErrorType.DIV_BY_ZERO, f: '=1/0' })).toBe(ErrorType.DIV_BY_ZERO);
-        expect(extractFormulaError({ v: ErrorType.NAME, f: '=S' })).toBe(ErrorType.NAME);
-        expect(extractFormulaError({ v: ErrorType.VALUE, f: '=SUM(A1)' })).toBe(ErrorType.VALUE);
-        expect(extractFormulaError({ v: ErrorType.NUM, f: '=SUM(A1)' })).toBe(ErrorType.NUM);
-        expect(extractFormulaError({ v: ErrorType.NA, f: '=SUM(A1)' })).toBe(ErrorType.NA);
-        expect(extractFormulaError({ v: ErrorType.CYCLE, f: '=SUM(A1)' })).toBe(ErrorType.CYCLE);
-        expect(extractFormulaError({ v: ErrorType.REF, f: '=SUM(A1)' })).toBe(ErrorType.REF);
-        expect(extractFormulaError({ v: ErrorType.SPILL, f: '=SUM(A1)' })).toBe(ErrorType.SPILL);
-        expect(extractFormulaError({ v: ErrorType.CALC, f: '=SUM(A1)' })).toBe(ErrorType.CALC);
-        expect(extractFormulaError({ v: ErrorType.ERROR, f: '=SUM(A1)' })).toBe(ErrorType.ERROR);
-        expect(extractFormulaError({ v: ErrorType.CONNECT, f: '=SUM(A1)' })).toBe(ErrorType.CONNECT);
-
-        expect(extractFormulaError({ v: ErrorType.NULL, f: '=SUM(A1)' })).toBe(ErrorType.NULL);
-        expect(extractFormulaError({ v: ErrorType.NULL, f: '=SUM(A1)', si: 'id1' })).toBe(ErrorType.NULL);
-        expect(extractFormulaError({ v: ErrorType.NULL, si: 'id1' })).toBe(ErrorType.NULL);
-        expect(extractFormulaError({ v: ErrorType.NULL })).toBeNull();
-
-        expect(extractFormulaError({ f: '=SUM(1)' })).toBeNull();
-        expect(extractFormulaError({ si: 'id1' })).toBeNull();
-        expect(extractFormulaError({ p: null })).toBeNull();
-        expect(extractFormulaError({ v: 'test' })).toBeNull();
-        expect(extractFormulaError({ v: 1 })).toBeNull();
-        expect(extractFormulaError({})).toBeNull();
-        expect(extractFormulaError(null)).toBeNull();
-        expect(extractFormulaError(undefined)).toBeNull();
-    });
-
     it('Function extractFormulaNumber', () => {
         const v = 0.07 / 0.1;
         expect(extractFormulaNumber({ v })).toBeNull();
